Follow-up fixes for commit c916cdf323b86aab132c1f2a4bb2cdf6a56873e2.
This commit itself is a follow-up for bug 11638, but with PostgreSQL
9.0 and its VACUUM FULL rewrite didn't cover all cases within the REL8_4_LTS
branch.

The various places fixes the call to vac_update_relstats() to honor
the current nested transaction state, wether its run within a transaction
or not. Since all of those places seem to be employed by VACUUM, we don't
need to bother about the current transaction context and simply pass false
to the function, afaics.