Update expected vaccum regression test result.
Since 64d024fc1426a996d5ffddb8a0056dc3c86dd5c2, the vacuum regression test is
failing.  Analysis has revealed that the real culprit was commit
5008ee54c33614b399b06f937dea3bed5e1dd34d.  Due to the different implementations
of VACUUM FULL in 8.4 compared to later branches, the reentrant problem fixed in
the commit is not triggered in 8.4, so the error in the expected test output is
not present.  Removing this additional error output fixes the test.